requirements.txt
alembic==1.9.3
anyio==3.6.2
asyncpg==0.27.0
click==8.1.3
fastapi==0.79.0
flake8==6.0.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
isort==5.12.0
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
orjson==3.4.1
pkg_resources==0.0.0
pycodestyle==2.10.0
pydantic==1.9.0
pyflakes==3.0.1
python-dotenv==0.21.1
sniffio==1.3.0
SQLAlchemy==2.0.3
starlette==0.19.1
typing_extensions==4.4.0
uvicorn==0.18.2
uvloop==0.14.0
zipp==3.13.0
